requirements.txt
alembic==0.8.9
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
bleach==1.4.2
blinker==1.4
click==6.7
colorama==0.3.7
convertdate==2.1.0
coverage==4.0.3
dateparser==0.5.1
decorator==4.0.10
dominate==2.2.0
enum34==1.1.6
ephem==3.7.6.0
flake8==2.5.4
Flask==0.10.1
Flask-Babel==0.11.1
Flask-Bootstrap==3.3.5.7
Flask-Bower==1.2.1
Flask-HTTPAuth==3.1.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Moment==0.5.1
Flask-OpenID==1.2.5
flask-paginate==0.4.5
Flask-Resize==0.8.1
Flask-Script==2.0.5
Flask-SocketIO==2.8.2
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14
flipflop==1.0
ForgeryPy==0.1
funcsigs==1.0.2
guess-language==0.2
html5lib==0.9999999
httpie==0.9.3
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
jdatetime==1.8.1
Jinja2==2.8
Mako==1.0.6
Markdown==2.6.6
MarkupSafe==0.23
mccabe==0.4.0
mock==2.0.0
nltk==3.2.2
olefile==0.44
pathlib2==2.1.0
pbr==1.10.0
pep8==1.7.0
pickleshare==0.7.4
pilkit==1.1.13
Pillow==4.0.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
pyflakes==1.0.0
Pygments==2.1.3
python-dateutil==2.6.0
python-editor==1.0.3
python-engineio==1.1.0
python-openid==2.2.5
python-socketio==1.6.2
pytz==2016.10
pyuploadcare==2.1.0
pyuploadcare-wtforms==0.5
ranking==0.3.1
regex==2016.12.27
requests==2.9.1
ruamel.ordereddict==0.4.9
ruamel.yaml==0.13.7
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.12
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
traitlets==4.3.1
typing==3.5.3.0
tzlocal==1.3
umalqurra==0.2
visitor==0.1.2
wcwidth==0.1.7
Werkzeug==0.11.5
Whoosh==2.7.4
win-unicode-console==0.5
WTForms==2.1
